requirements.txt
attrs==21.4.0
black==22.3.0
bleach==5.0.0
build==0.8.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
docutils==0.18.1
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
keyring==23.6.0
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pep517==0.12.0
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
pytest-mock==3.7.0
python-dotenv==0.20.0
PyYAML==6.0
readme-renderer==35.0
requests==2.28.0
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.4
six==1.16.0
tomli==2.0.1
twine==4.0.1
urllib3==1.26.9
vcrpy==4.1.1
webencodings==0.5.1
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
